requirements.txt
fastapi==0.121.2
uvicorn==0.34.0
supabase==2.15.0
python-dotenv==1.0.1
langchain==0.3.20
langchain-openai==0.3.12
pydantic==2.12.4